refactor test structure Didn't change any of the actual tests, just organized it
