+++ test-LaPlaya/CPComments.conf
@@ -2,15 +2,15 @@
 # Main section
 #
 [main]
-title=Critical Path - Show Projects
-description=Load test the potential critical path involved in showing projects.
-#url=http://ec2-54-69-120-176.us-west-2.compute.amazonaws.com/
-url=http://localhost:3000/
+title=Critical Path - Comments
+description=Load test the potential critical paths involved with viewing or creating comments
+#url=http://ec2-54-69-120-176.us-west-2.compute.amazonaws.com
+url=http://localhost:3000
 
 # ------------------------------------------------------------
 # Configuration for unit test mode fl-run-test
 #
-# run with 'fl-run-test test_CPShowProjects.py'
+# run with 'fl-run-test test_CPComments.py'
 #
 [ftest]
 # log_to destination =
@@ -19,10 +19,10 @@ url=http://localhost:3000/
 log_to = console file
 
 # log_path = path and file name to store log file
-log_path = cp-createuser.log
+log_path = cp-comments.log
 
 # result_path = path to store the xml result file
-result_path = cp-createuser.xml
+result_path = cp-comments.xml
 
 # sleeptime_min = minimum amount of time in seconds to sleep between requests
 #                 to the host
@@ -35,7 +35,7 @@ sleep_time_max = 0.5
 # ------------------------------------------------------------
 # Configuration for bench mode fl-run-bench
 #
-# run with 'fl-run-bench test_CPShowProjects.py CPShowProjects.test_critical_path'
+# run with 'fl-run-bench test_CPComments.py CPComments.test_readonly_view_comment'
 #
 [bench]
 description=Load the projects page with many concurrent users
@@ -57,7 +57,7 @@ cycle_time = 1
 
 # same keys as in [ftest] section
 log_to = 
-log_path = cp-createuser-bench.log
-result_path = cp-createuser-bench.xml
+log_path = cp-comments-bench.log
+result_path = cp-comments-bench.xml
 sleep_time_min = 0
 sleep_time_max = 0.5
