+++ test-LaPlaya/CPCreateUser.conf
@@ -1,8 +1,8 @@
 [main]
 title=Critical Path - Create User
 description=Load test the potential critical path involved in creating a user.
-#url=http://ec2-54-69-120-176.us-west-2.compute.amazonaws.com/
-url=http://localhost:3000/
+#url=http://ec2-54-69-120-176.us-west-2.compute.amazonaws.com
+url=http://localhost:3000
 
 #[test_first]
 #description=Access the main URL %(nb_time) times
