+++ test-LaPlaya/CPShowProjects.conf
@@ -1,0 +1,63 @@
+# ------------------------------------------------------------
+# Main section
+#
+[main]
+title=Critical Path - Show Projects
+description=Load test the potential critical path involved in showing projects.
+#url=http://ec2-54-69-120-176.us-west-2.compute.amazonaws.com
+url=http://localhost:3000
+
+# ------------------------------------------------------------
+# Configuration for unit test mode fl-run-test
+#
+# run with 'fl-run-test test_CPShowProjects.py'
+#
+[ftest]
+# log_to destination =
+# console - to the screen
+# file - to a file
+log_to = console file
+
+# log_path = path and file name to store log file
+log_path = cp-showprojects.log
+
+# result_path = path to store the xml result file
+result_path = cp-showprojects.xml
+
+# sleeptime_min = minimum amount of time in seconds to sleep between requests
+#                 to the host
+sleep_time_min = 0
+
+# sleeptime_max = maximum amount of time in seconds to sleep between requests
+#                 to the host
+sleep_time_max = 0.5
+
+# ------------------------------------------------------------
+# Configuration for bench mode fl-run-bench
+#
+# run with 'fl-run-bench test_CPShowProjects.py CPShowProjects.test_readonly_view_projects'
+#
+[bench]
+description=Load the projects page with many concurrent users
+
+# cycles = list of cycles with their number of concurrent users
+cycles = 2:4:6
+
+# duration = duration of a cycle in seconds
+duration = 10
+
+# startup_delay = time to wait between starting-up threads in seconds
+startup_delay = 0.01
+
+# sleep_time = time to wait between test in seconds
+sleep_time = 0.01
+
+# cycle_time = time to wait between cycle in seconds
+cycle_time = 1
+
+# same keys as in [ftest] section
+log_to = 
+log_path = cp-showprojects-bench.log
+result_path = cp-showprojects-bench.xml
+sleep_time_min = 0
+sleep_time_max = 0.5
